pyproject.toml
[project]
name = "epaper-display-panel"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gpiozero>=2.0",
    "lgpio>=0.2",
    "pillow>=11.0",
    "requests>=2.32",
    "spidev>=3.6",
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "I", "F"]

[dependency-groups]
dev = [
    "pre-commit>=4.0",
    "ruff>=0.8",
]
